Bolt/requirements.txt
nicegui
pyserial
